test(core): split define alias tests
